+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### v4.37.0 Features
 
+- `[Toolbar Searchfield]` Fixed a bug where the toolbar searchfield were unable to focused when tabbing through the page. ([#4683](https://github.com/infor-design/enterprise/issues/4683))
 - `[Toolbar Searchfield]` Fixed a bug where the search bar were showing extra outline when focused. ([#4682](https://github.com/infor-design/enterprise/issues/4682))
 
 ### v4.37.0 Fixes